dbeaver/pro#8126 create temp folders in cli
