+++ plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/core/DesktopPlatform.java
@@ -1,6 +1,6 @@
 /*
  * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2025 DBeaver Corp and others
+ * Copyright (C) 2010-2026 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -33,37 +33,25 @@ import org.jkiss.dbeaver.model.navigator.DesktopNavigatorModel;
 import org.jkiss.dbeaver.model.preferences.DBPPreferenceStore;
 import org.jkiss.dbeaver.model.qm.QMRegistry;
 import org.jkiss.dbeaver.model.qm.QMUtils;
-import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
 import org.jkiss.dbeaver.model.runtime.LoggingProgressMonitor;
 import org.jkiss.dbeaver.model.runtime.features.DBRFeatureRegistry;
-import org.jkiss.dbeaver.registry.BasePlatformImpl;
+import org.jkiss.dbeaver.registry.BaseLocalPlatform;
 import org.jkiss.dbeaver.registry.DataSourceProviderRegistry;
 import org.jkiss.dbeaver.registry.GlobalEventManagerImpl;
 import org.jkiss.dbeaver.runtime.SecurityProviderUtils;
 import org.jkiss.dbeaver.runtime.qm.QMLogFileWriter;
 import org.jkiss.dbeaver.runtime.qm.QMRegistryImpl;
 import org.jkiss.dbeaver.utils.ContentUtils;
-import org.jkiss.dbeaver.utils.GeneralUtils;
 import org.jkiss.dbeaver.utils.RuntimeUtils;
-import org.jkiss.dbeaver.utils.SystemVariablesResolver;
-import org.jkiss.utils.CommonUtils;
-import org.jkiss.utils.StandardConstants;
-
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 
 /**
  * DesktopPlatform
  */
-public class DesktopPlatform extends BasePlatformImpl implements DBPPlatformDesktop {
+public class DesktopPlatform extends BaseLocalPlatform implements DBPPlatformDesktop {
 
     // The plug-in ID
     public static final String PLUGIN_ID = "org.jkiss.dbeaver.core"; //$NON-NLS-1$
 
-    private static final String TEMP_PROJECT_NAME = ".dbeaver-temp"; //$NON-NLS-1$
     private static final String DBEAVER_CONFIG_FOLDER = "settings";
 
     private static final Log log = Log.getLog(DesktopPlatform.class);
@@ -72,7 +60,6 @@ public class DesktopPlatform extends BasePlatformImpl implements DBPPlatformDesk
 
     private static volatile boolean isClosing = false;
 
-    private Path tempFolder;
     private DBPWorkspaceDesktop workspace;
     private QMRegistryImpl queryManager;
     private QMLogFileWriter qmLogWriter;
@@ -253,64 +240,6 @@ public class DesktopPlatform extends BasePlatformImpl implements DBPPlatformDesk
 
     public void setWorkbenchStarted(boolean started) {
         this.workbenchStarted = started;
-    }
-
-    @NotNull
-    public Path getTempFolder(@NotNull DBRProgressMonitor monitor, @NotNull String name) {
-        if (tempFolder == null) {
-            // Make temp folder
-            try {
-                String tempFolderPath = System.getProperty("dbeaver.io.tmpdir");
-                if (!CommonUtils.isEmpty(tempFolderPath)) {
-                    tempFolderPath = GeneralUtils.replaceVariables(tempFolderPath, new SystemVariablesResolver());
-
-                    File dbTempFolder = new File(tempFolderPath);
-                    if (!dbTempFolder.mkdirs()) {
-                        throw new IOException("Can't create temp directory '" + dbTempFolder.getAbsolutePath() + "'");
-                    }
-                } else {
-                    tempFolderPath = System.getProperty(StandardConstants.ENV_TMP_DIR);
-                }
-                monitor.subTask("Create temp folder '" + tempFolderPath + "'");
-                Path tmpFolder = Paths.get(tempFolderPath);
-                if (!Files.exists(tmpFolder)) {
-                    log.debug("Create global temp folder '" + tmpFolder + "'");
-                    Files.createDirectories(tmpFolder);
-                }
-                tempFolder = Files.createTempDirectory(tmpFolder, TEMP_PROJECT_NAME);
-            } catch (IOException e) {
-                final String sysTempFolder = System.getProperty(StandardConstants.ENV_TMP_DIR);
-                if (!CommonUtils.isEmpty(sysTempFolder)) {
-                    tempFolder = Path.of(sysTempFolder).resolve(TEMP_PROJECT_NAME);
-                    if (!Files.exists(tempFolder)) {
-                        try {
-                            Files.createDirectories(tempFolder);
-                        } catch (IOException ex) {
-                            final String sysUserFolder = System.getProperty(StandardConstants.ENV_USER_HOME);
-                            if (!CommonUtils.isEmpty(sysUserFolder)) {
-                                tempFolder = Path.of(sysUserFolder).resolve(TEMP_PROJECT_NAME);
-                                if (!Files.exists(tempFolder)) {
-                                    try {
-                                        Files.createDirectories(tempFolder);
-                                    } catch (IOException exc) {
-                                        tempFolder = Path.of(TEMP_PROJECT_NAME);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        Path localTemp = name == null ? tempFolder : tempFolder.resolve(name);
-        if (!Files.exists(localTemp)) {
-            try {
-                Files.createDirectories(localTemp);
-            } catch (IOException e) {
-                log.error("Can't create temp directory " + localTemp, e);
-            }
-        }
-        return localTemp;
     }
 
     @Override
